Expose networking.privateIPv4 to backend
the rationale behind this change is to allow
some backends refer user specified IP address
and use it for ssh connection.

For example, imagine a usecase that user wants
the machine to run with a pre-determined statically
configured IP address instead of relying on randomly
assigned an dynamic address from DHCP in libvirtd or vbox.